RR-1198 - Move 2 more Review Schedule functions to ReviewScheduleService
