+++ domain/learningandworkprogress/src/test/kotlin/uk/gov/justice/digital/hmpps/domain/learningandworkprogress/review/service/ReviewServiceTest.kt
@@ -1,7 +1,6 @@
 package uk.gov.justice.digital.hmpps.domain.learningandworkprogress.review.service
 
 import org.assertj.core.api.Assertions.assertThat
-import org.assertj.core.api.Assertions.catchThrowableOfType
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
@@ -11,9 +10,7 @@ import org.mockito.junit.jupiter.MockitoExtension
 import org.mockito.kotlin.any
 import org.mockito.kotlin.given
 import org.mockito.kotlin.verify
-import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.review.ReviewScheduleNotFoundException
 import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.review.aValidCompletedReview
-import uk.gov.justice.digital.hmpps.domain.learningandworkprogress.review.aValidReviewSchedule
 import uk.gov.justice.digital.hmpps.domain.randomValidPrisonNumber
 
 @ExtendWith(MockitoExtension::class)
@@ -30,41 +27,11 @@ class ReviewServiceTest {
   @Mock
   private lateinit var reviewSchedulePersistenceAdapter: ReviewSchedulePersistenceAdapter
 
+  @Mock
+  private lateinit var reviewScheduleService: ReviewScheduleService
+
   companion object {
     private val PRISON_NUMBER = randomValidPrisonNumber()
-  }
-
-  @Nested
-  inner class GetActiveReviewScheduleForPrisoner {
-    @Test
-    fun `should get active review schedule for prisoner`() {
-      // Given
-      val expected = aValidReviewSchedule()
-      given(reviewSchedulePersistenceAdapter.getActiveReviewSchedule(any())).willReturn(expected)
-
-      // When
-      val actual = service.getActiveReviewScheduleForPrisoner(PRISON_NUMBER)
-
-      // Then
-      assertThat(actual).isEqualTo(expected)
-      verify(reviewSchedulePersistenceAdapter).getActiveReviewSchedule(PRISON_NUMBER)
-    }
-
-    @Test
-    fun `should fail to get review schedule for prisoner given review schedule does not exist`() {
-      // Given
-      given(reviewSchedulePersistenceAdapter.getActiveReviewSchedule(any())).willReturn(null)
-
-      // When
-      val exception = catchThrowableOfType(ReviewScheduleNotFoundException::class.java) {
-        service.getActiveReviewScheduleForPrisoner(PRISON_NUMBER)
-      }
-
-      // Then
-      assertThat(exception).hasMessage("Review Schedule not found for prisoner [$PRISON_NUMBER]")
-      assertThat(exception.prisonNumber).isEqualTo(PRISON_NUMBER)
-      verify(reviewSchedulePersistenceAdapter).getActiveReviewSchedule(PRISON_NUMBER)
-    }
   }
 
   @Nested
